Add polymer element: lmv-button-group
Notice that there used to be a component named like this, but it was
recently renamed to lmv-button-menu

lmv-button-group is a container that makes sure that only one lmv-button
is ever active at the same time. It also supports having zero buttons
active at a time. Never two or more.

TODO: Add unit tests